+++ src/Controllers/Private/Permissions.php
@@ -2,6 +2,7 @@
 
 namespace Bayfront\BonesService\Api\Controllers\Private;
 
+use Bayfront\ArrayHelpers\Arr;
 use Bayfront\BonesService\Api\ApiService;
 use Bayfront\BonesService\Api\Controllers\Abstracts\PrivateApiController;
 use Bayfront\BonesService\Api\Exceptions\ApiServiceException;
@@ -13,8 +14,14 @@ use Bayfront\BonesService\Api\Exceptions\Http\TooManyRequestsException;
 use Bayfront\BonesService\Api\Interfaces\CrudControllerInterface;
 use Bayfront\BonesService\Api\Schemas\PermissionCollection;
 use Bayfront\BonesService\Api\Schemas\PermissionResource;
+use Bayfront\BonesService\Api\Schemas\TenantCollection;
 use Bayfront\BonesService\Api\Traits\UsesResourceModel;
+use Bayfront\BonesService\Orm\Exceptions\InvalidRequestException;
+use Bayfront\BonesService\Orm\Exceptions\UnexpectedException;
+use Bayfront\BonesService\Orm\Utilities\Parsers\QueryParser;
 use Bayfront\BonesService\Rbac\Models\PermissionsModel;
+use Bayfront\BonesService\Rbac\Models\TenantPermissionsModel;
+use Bayfront\BonesService\Rbac\Models\TenantsModel;
 
 class Permissions extends PrivateApiController implements CrudControllerInterface
 {
@@ -151,6 +158,74 @@ class Permissions extends PrivateApiController implements CrudControllerInterfac
         $this->deleteResource($this->permissionsModel, $params['id']);
 
         $this->respond(204);
+
+    }
+
+    /**
+     * List tenants with permission.
+     *
+     * @param array $params
+     * @return void
+     * @throws ApiServiceException
+     * @throws BadRequestException
+     * @throws ForbiddenException
+     * @throws NotFoundException
+     */
+    public function listTenants(array $params): void
+    {
+
+        $this->validatePath($params, [
+            'id' => 'required|uuid'
+        ]);
+
+        $this->validateQuery($this->getQueryParserRules());
+
+        $this->validateIsAdmin($this->user);
+
+        // Ensure permission exists
+
+        if (!$this->resourceExists($this->permissionsModel, $params['id'])) {
+            throw new NotFoundException();
+        }
+
+        // Get array of permission ID's
+
+        $tenantPermissionsModel = new TenantPermissionsModel($this->rbacService);
+
+        try {
+
+            $tenantCollection = $tenantPermissionsModel->list(new QueryParser([
+                'fields' => 'tenant',
+                'filter' => [
+                    [
+                        'permission' => [
+                            'eq' => $params['id']
+                        ]
+                    ]
+                ]
+            ]), true);
+
+        } catch (InvalidRequestException|UnexpectedException $e) {
+            throw new ApiServiceException($e->getMessage());
+        }
+
+        $tenant_ids = Arr::pluck($tenantCollection->list(), 'tenant');
+
+        // List tenants
+
+        $tenantsModel = new TenantsModel($this->rbacService);
+
+        $query_filter = [
+            [
+                'id' => [
+                    'in' => implode(',', $tenant_ids)
+                ]
+            ]
+        ];
+
+        $collection = $this->listResources($tenantsModel, $query_filter);
+
+        $this->respond(200, TenantCollection::create($collection['list'], $collection['config']));
 
     }